create input for plan section, pass in the correct props
