+++ src/components/FormInput/PlanInput.tsx
@@ -1,10 +1,18 @@
-const PlanInput = () => {
+import { UseControllerProps, useController } from 'react-hook-form'
+import { PlanForm } from '../Form/PlanSection/PlanSection.tsx'
+
+type PlanInputProps = UseControllerProps<PlanForm> & {
+  value: PlanForm['plan']
+}
+
+const PlanInput = (props: PlanInputProps) => {
+  const { field, fieldState } = useController(props)
   return (
     <div>
       <label>
         <img src="imageSource" alt="plan icon" />
-        <input type="radio" name="plan" value="plan" />
-        <p>plan name</p>
+        <input {...field} value={props.value} name={props.name} />
+        <p>{props.name}</p>
         <p>plan price</p>
       </label>
     </div>
